+++ frontend/src/components/features/account/auth/forms/LoginForm.tsx
@@ -1,17 +1,15 @@
 'use client'
 
 import type { FieldValues, SubmitHandler } from 'react-hook-form'
-
 import type { z } from 'zod'
-
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
+import { Button } from '~/components/ui/button'
+import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '~/components/ui/form'
+import { Input } from '~/components/ui/input'
 import { useAuth } from '~/hooks/useAuth'
 import { logger } from '~/lib/logger'
 import { loginSchema } from '~/schemas/login.schema'
-import { Button } from '../../../../ui/button'
-import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from './Form'
-import { Input } from '../../../../ui/input'
 
 interface LoginFormProps {
   onSuccess?: () => void
